REGRESSION(256902@main): [WPE] Broke context-menu signal https://bugs.webkit.org/show_bug.cgi?id=248243
Reviewed by Carlos Garcia Campos.

256902@main accidentally broke WPE's WebKitWebView::context-menu signal.
After that commit, the signal declaration now has two parameters, but
the signal emission has three parameters. Instead, the original API
needs to have three parameters, while the new API can have just two
parameters. This means WebKitWebViewWPE needs to be split into multiple
files.

* Source/WebKit/PlatformWPE.cmake:
* Source/WebKit/SourcesWPE.txt:
* Source/WebKit/UIProcess/API/glib/WebKitWebView.cpp:
(webkitWebViewPopulateContextMenu):
* Source/WebKit/UIProcess/API/wpe/WebKitWebViewWPE.cpp:
(createContextMenuSignal): Deleted.
* Source/WebKit/UIProcess/API/wpe/WebKitWebViewWPE1.cpp: Added.
(createContextMenuSignal):
* Source/WebKit/UIProcess/API/wpe/WebKitWebViewWPE2.cpp: Added.
(createContextMenuSignal):

Canonical link: https://commits.webkit.org/256991@main
